+++ src/liff-stub.ts
@@ -1,5 +1,8 @@
 import { Liff } from './types';
 
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+type LiffStub = Liff & { [key in Exclude<string, keyof Liff>]: any };
+
 // login status
 let loggedIn = false;
 
@@ -8,7 +11,7 @@ const permanentLink: Liff['permanentLink'] = {
   setExtraQueryParam: () => {},
 };
 
-export const liffStub: Liff = {
+export const liffStub: LiffStub = {
   init: async () => {},
   getOS: () => 'web',
   getVersion: () => '2.1.3',