+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 ## Features
 ## Bugfixes
+* [#107](https://github.com/epaew/react-liff/pull/107) Fixup type error with `@line/liff@2.4.0`.
+
 ## Others
 
 # 0.5.0